chore(upsell): move DescriptionSerializer to components/UpsellDescriptionSerializer
