+++ packages/sanity/src/core/components/index.ts
@@ -30,3 +30,4 @@ export * from './documentStatusIndicator'
 export * from './loadingBlock'
 export * from './documentStatus'
 export * from './contextMenuButton'
+export * from './upsellDescriptionSerializer'
