+++ packages/sanity/src/core/components/upsellDescriptionSerializer/helpers.ts
