+++ packages/sanity/src/core/components/upsellDescriptionSerializer/index.ts
@@ -1,0 +1,1 @@
+export * from './UpsellDescriptionSerializer'
