+++ packages/sanity/src/core/studio/components/navbar/index.ts
@@ -2,4 +2,3 @@ export * from './StudioLogo'
 export * from './StudioNavbar'
 export * from './search'
 export * from './tools'
-export * from './free-trial/DescriptionSerializer'
